reistall update

+++ .config/nvim/lua/cfg/fterm.lua
@@ -1,9 +1,8 @@
 require("FTerm").setup({
-	border = "double",
+	border = "single",
 	dimensions = {
 		height = 0.9,
 		width = 0.9,
 	},
---	blend = 13,
+	--	blend = 13,
 })
-
